pyproject.toml
[tool.poetry]
name = "horseracingai"
version = "1.0.0"
description = "High-precision horse racing prediction system"
authors = ["HorseRacingAI Team"]
readme = "README.md"
packages = [{include = "services"}, {include = "database"}]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.11.1"
selenium = "^4.8.0"
pandas = "^1.5.3"
numpy = "^1.24.2"
scikit-learn = "^1.2.2"
lightgbm = "^3.3.5"
tensorflow = "^2.12.0"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
sqlalchemy = "^2.0.0"
alembic = "^1.10.2"
celery = {extras = ["redis"], version = "^5.2.7"}
redis = "^4.5.1"
pydantic = "^1.10.7"
psycopg2-binary = "^2.9.5"
pymongo = "^4.3.3"
minio = "^7.1.14"
httpx = "^0.23.3"
python-dotenv = "^1.0.0"
loguru = "^0.6.0"
pytest = "^7.2.2"
pytest-asyncio = "^0.21.0"
black = "^23.1.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.11.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"